Conflicts:
	praktiskprove/knutepunkt/src/main/java/no/difi/meldingsutveksling/noarkexchange/SendMessageTemplate.java

+++ praktiskprove/knutepunkt/src/main/resources/db/schema/ddl.sql
@@ -2,7 +2,7 @@ CREATE TABLE EVENT_LOG (
                 UUID VARCHAR(36), 
                 SENDER VARCHAR(9), 
                 RECEIVER VARCHAR(9), 
-                EVENT_TIMESTAMP INT, 
+                EVENT_TIMESTAMP BIGINT,
                 STATE VARCHAR(50), 
                 ERROR_MESSAGE  LONG VARCHAR,
                 MESSAGE LONG VARCHAR,
